Initial commit for mviews as they currently are
The liquibase sql script just regenerates the mviews as they currently are.
The create scripts are generated by pgAdmin while pointing at the prdg
DB.

Commit also, includes pgTap docker compose file and tests which pass against
current versions of the mviews.
Also, there is an override for the now() function which may be useful
later.