FSR-595: Fixed incorrect sequence name for station_threshold_id
The table and sequence was renamed in an earlier commit but this was
missed.

Looks like the script was dropping and recreating the sequence correctly
but was actually using a preexisting sequence with the old name.

This only became apparent when running the script in an environment
where the original version had never been run (i.e. in the tst
environment)